requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.3
decorator==5.1.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
